// src/app/shared/schemas/community/member-bio.ts
export class MemberBio {
  id?: number;
  user_id?: number;
  username?: string;
  first_name?: string;
  last_name?: string;
  member_description?: string;
  image_id?: string;
  company_name?: string;
  external_url?: string;
  status?: string;
  deleted?: string;
  listedItems?: number;
  totalItemsSold?: number;
  country?: string;
  role_id?: number;
  cover_image?: string;
  sort?: number;
  constructor() {}
}